Address [-Wsign-compare] warning.

+++ moveit_core/distance_field/src/propagation_distance_field.cpp
@@ -663,9 +663,9 @@ bool PropagationDistanceField::writeToStream(std::ostream& os) const
   out.push(boost::iostreams::zlib_compressor());
   out.push(os);
 
-  for(unsigned int x = 0; x < getXNumCells(); x++) {
-    for(unsigned int y = 0; y < getYNumCells(); y++) {
-      for(unsigned int z = 0; z < getZNumCells(); z+=8) {
+  for(unsigned int x = 0; x < static_cast<unsigned int>(getXNumCells()); x++) {
+    for(unsigned int y = 0; y < static_cast<unsigned int>(getYNumCells()); y++) {
+      for(unsigned int z = 0; z < static_cast<unsigned int>(getZNumCells()); z+=8) {
         std::bitset<8> bs(0);
         unsigned int zv = std::min((unsigned int)8, getZNumCells()-z);
         for(unsigned int zi = 0; zi < zv; zi++) {
@@ -731,9 +731,9 @@ bool PropagationDistanceField::readFromStream(std::istream& is)
   //std::cout << "Nums " << getXNumCells() << " " << getYNumCells() << " " << getZNumCells() << std::endl;
 
   std::vector<Eigen::Vector3i> obs_points;
-  for(unsigned int x = 0; x < getXNumCells(); x++) {
-    for(unsigned int y = 0; y < getYNumCells(); y++) {
-      for(unsigned int z = 0; z < getZNumCells(); z+=8) {
+  for(unsigned int x = 0; x < static_cast<unsigned int>(getXNumCells()); x++) {
+    for(unsigned int y = 0; y < static_cast<unsigned int>(getYNumCells()); y++) {
+      for(unsigned int z = 0; z < static_cast<unsigned int>(getZNumCells()); z+=8) {
         char inchar;
         if(!in.good()) {
           return false;
